changed longs to ints

+++ payment-project/src/main/java/com/progressoft/jip/payment/account/AccountDTO.java
@@ -20,7 +20,7 @@ public interface AccountDTO extends DTO {
 
 	AccountStatus getAccountStatus();
 
-	long getIbanId();
+	int getIbanId();
 
 	public enum AccountStatus {
